requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.30.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
pydantic==2.12.5
SQLAlchemy==2.0.48
psycopg[binary]==3.1.18
redis==5.0.3
websockets==12.0
requests==2.31.0